requirements.txt
# HTTP library for API requests
requests>=2.31.0,<3.0.0

# YAML parser
PyYAML>=6.0.1,<7.0.0

# Data manipulation library
pandas>=2.0.0,<3.0.0
